Fixed failure to install on systems where locale is not UTF-8
In Python 3 open() defaults to reading in text mode in the default locale
for the system. This causes an exception if the default locale encoding
is ASCII, since the README.rst file is UTF-8 and not valid ASCII.
The encoding can't be specified in the call to open() since this would
not be valid Python 2.